requirements.txt
Django
celery==4.1.0
djangorestframework